RaspberryPi/requirements.txt
PyQt5
pyqtgraph
paho-mqtt
mysql-connector-python